requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
Jinja2
python-multipart
